It's now EntitySaddleEvent
