--- src/main/java/org/bukkit/event/entity/PigSaddleEvent.java
@@ -1,34 +1,0 @@
-package org.bukkit.event.entity;
-
-import org.bukkit.entity.Pig;
-import org.bukkit.event.Cancellable;
-import org.bukkit.event.HandlerList;
-
-/**
- * Called when a pig is saddled
- */
-public class PigSaddleEvent extends EntityEvent implements Cancellable {
-    private static final HandlerList handlers = new HandlerList();
-    private boolean cancel;
-
-    public PigSaddleEvent(final Pig pig) {
-        super(pig);
-    }
-
-    public boolean isCancelled() {
-        return cancel;
-    }
-
-    public void setCancelled(boolean cancel) {
-        this.cancel = cancel;
-    }
-
-    @Override
-    public HandlerList getHandlers() {
-        return handlers;
-    }
-
-    public static HandlerList getHandlerList() {
-        return handlers;
-    }
-}
